# doc/openapi 3.yml
openapi: 3.0.0
info:
  title: API TRUFO
  description: API para el registro y consulta de avistamientos de OVNIs.
  version: "1.0.0"
security:
  - ApiKeyAuth: []
  

paths:
  /sightings:
    get:
      security:
        - ApiKeyAuth: []
      summary: Obtiene una lista de avistamientos
      description: Devuelve una lista de registros de avistamientos. Se pueden aplicar filtros por ciudad, estado, país, forma y rango de fechas, 
        además de usar paginación.
      parameters:
        - name: limit
          in: query
          description: Número de resultados por página (por defecto 10)
          required: false
          schema:
            type: integer
            default: 10
        - name: shape
          in: query
          description: Filtrar por forma del OVNI
          required: false
          schema:
            type: string
        - name: from_datetime
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to_datetime
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: next
          in: query
          description: URL para la siguiente página de resultados
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de avistamientos correspondiente a los filtros aplocados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sightings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sightings_reduced'
                  next:
                    description: URL para la siguiente página de resultados
                    type: string
        '400':
          description: Error al buscar avistamientos
        '500':
          description: Error interno del servidor

    post:
      security:
        - ApiKeyAuth: []
      summary: Crea un nuevo avistamiento
      description: Añade un nuevo OVNI a la DB
      requestBody:
        description: Objeto con la información del avistamiento a crear
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sighting'
      responses:
        '201':
          description: Avistamiento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sighting'
        '400':
          description: Datos incorrectos

  /sightings/{sightingId}:
    get:
      security:
        - ApiKeyAuth: []
      summary: Obtiene un avistamiento específico
      description: Devuelve la información de un avistamiento por id
      parameters:
        - name: sightingId
          in: path
          description: Id del avistamiento
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Avistamiento obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sighting'
        '404':
          description: Avistamiento no encontrado

    put:
      security:
        - ApiKeyAuth: []
      summary: Actualiza un avistamiento existente
      description: Modifica los datos de un avistamiento especifico
      parameters:
        - name: sightingId
          in: path
          description: Id del avistamiento
          required: true
          schema:
            type: string
      requestBody:
        description: Objeto con los datos actualizados del avistamiento.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sighting'
      responses:
        '200':
          description: Avistamiento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sighting'
        '400':
          description: Datos incorrectos, no actualizado
        '404':
          description: Avistamiento no encontrado

    delete:
      security:
        - ApiKeyAuth: []
      summary: Elimina un avistamiento
      description: Elimina el registro de un avistamiento identificado por su ID.
      parameters:
        - name: sightingId
          in: path
          description: Id del avistamiento
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Avistamiento eliminado exitosamente
        '404':
          description: Avistamiento no encontrado
        '500':
          description: Error al eliminar el avistamiento

  /sightings/user:
    get:
      security:
        - ApiKeyAuth: []
      summary: Obtiene los avistamientos de un usuario concreto
      description: Devuelve una lista de avistamientos filtrados por el correo electrónico del usuario.
      parameters:
        - name: user_email
          in: query
          description: Correo electrónico del usuario
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de avistamientos del usuario
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sighting'
        '404':
          description: No se encontraron avistamientos para ese usuario
        '500':
          description: Error interno del servidor  
          
  /locations/:latitud/:longitud:
    get:
      security:
        - ApiKeyAuth: []
      summary: Obtiene la ubicación de un avistamiento a partir de sus coordenadas
      description: Devuelve una lista de ubicaciones donde han ocurrido avistamientos.
      responses:
        '200':
          description: Lista de ubicaciones obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/location'
        '404':
          description: No se encontraron resultados para las coordenadas proporcionadas
        '500':
          description: Error al conectar con el servicio de geolocalización
 
  
  /people:
    get:
      security:
        - ApiKeyAuth: []
      summary: Obtiene una lista de personas
      description: Devuelve una lista de personas que han reportado avistamientos.
      responses:
        '200':
          description: Lista de personas obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/People'
        '500':
          description: Error interno del servidor
    post:
      security:
        - ApiKeyAuth: []
      summary: Crea una nueva persona
      description: Añade una nueva persona a la base de datos.
      requestBody:
        description: Objeto con la información de la persona a crear
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/People'
      responses:
        '201':
          description: Persona creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/People'
        '400':
          description: Datos incorrectos
      
    

components:
  schemas:
    Sighting:
      type: object
      description: Registro de un avistamiento de OVNIs.
      properties:
        id:
          type: string
          description: Identificador único del avistamiento.
          example: "1"
        datetime:
          type: string
          format: date-time
          description: Fecha y hora del avistamiento.
          example: "1949-10-10T20:30:00Z" 
        shape:
          type: string
          description: Forma del objeto avistado.
          example: "cylinder"
        "duration (seconds)":
          type: integer
          description: Duración del avistamiento en segundos.
          example: 2700
        comments:
          type: string
          description: Comentarios o detalles adicionales sobre el avistamiento.
          example: "This event took place in early fall after a Boy Scout meeting..."
        date_posted:
          type: string
          format: date-time
          description: Fecha en que se publicó el reporte.
          example: "1949-11-10T20:30:00Z"
        latitude:
          type: number
          format: float
          description: Latitud del lugar donde ocurrió el avistamiento.
          example: 29.8830556
        longitude:
          type: number
          format: float
          description: Longitud del lugar donde ocurrió el avistamiento.
          example: -97.9411111
        user_email:
          type: string
          description: Correo electrónico de la persona que ha registrado el avistamiento.
          example: "example@email.com"
      required:
        - datetime
        - shape
        - "duration (seconds)"
        - comments
        - date_posted
        - latitude
        - longitude
        - user_email
    location:
      type: object
      description: Ubicación del avistamiento.
      properties:
        pais:
          type: string
          description: País donde ocurrió el avistamiento.
          example: "United States"
        estado:
          type: string
          description: Estado donde ocurrió el avistamiento.
          example: "Texas"
        latitud:
          type: number
          format: float
          description: Latitud del lugar donde ocurrió el avistamiento.
          example: 29.8830556
        longitud:
          type: number
          format: float
          description: Longitud del lugar donde ocurrió el avistamiento.
          example: -97.9411111
      required:
        -latitude
        -longitude
    People:
      type: object
      description: Persona que ha registrado un avistamiento
      properties:
        FirstName:
          type: string
          description: Nombre de la persona que ha registrado el avistamiento.
          example: "John"
        LastName:
          type: string
          description: Apellido de la persona que ha registrado el avistamiento.
          example: "Doe"
        Company:
          type: string
          description: Empresa donde trabaja la persona que ha registrado el avistamiento.
          example: "Tech Corp"
        Phone:
          type: string
          description: Número de teléfono de la persona que ha registrado el avistamiento.
          example: "+1-234-567-8900"
        Email:
          type: string
          description: Correo electrónico de la persona que ha registrado el avistamiento.
          example: "urangel@espinoza-francis.net"
      required:
        - FirstName
        - LastName
        - Company
        - Phone
        - Email
    Sightings_reduced:
      type: object
      description: Resumen de avistamientos.
      properties:
        _id:
          type: string
          description: Identificador único del avistamiento.
          example: "6829c355a75ae84c71c323d3"
        shape:
          type: string
          description: Forma del objeto avistado.
          example: "cylinder"
        location:
          type: string
          description: Url al endpoint donde obtener información de la ubicación del avistamiento.
          example: "http://localhost:3000/location/29.8830556/-97.9411111"
        UserEmail:
          type: string
          description: Correo electrónico de la persona que ha registrado el avistamiento.
          example: "urangel@espinoza-francis.net"
        
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clave de API para autenticar las solicitudes.
servers:
  - url: http://localhost:3000